add styles to authors list, combine lists loader to 1 independent
