+++ client/src/components/AuthorsList/AuthorImg/index.js
@@ -1,0 +1,3 @@
+import AuthorImg from './AuthorImg';
+
+export default AuthorImg;
