--- client/src/components/AuthorsList/AuthorsLoader/AuthorsLoader.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AuthorsLoader = () => {
-  return <p>Loading data</p>;
-};
-
-export default AuthorsLoader;
