--- client/src/components/AuthorsList/AuthorsLoader/index.js
@@ -1,3 +1,0 @@
-import AuthorsLoader from './AuthorsLoader';
-
-export default AuthorsLoader;
