--- client/src/components/CatalogList/CatalogLoader/index.js
@@ -1,3 +1,0 @@
-import CatalogLoader from './CatalogLoader';
-
-export default CatalogLoader;
