+++ client/src/components/ListLoader/index.js
@@ -1,0 +1,3 @@
+import ListLoader from './ListLoader';
+
+export default ListLoader;
